refactor: split damage resolution into focused modules
