+++ src/utils/gameplay/damage/constants.ts
@@ -1,0 +1,39 @@
+import { CombatLocation } from '@/types/gameplay';
+
+export const PILOT_DEATH_WOUND_THRESHOLD = 6;
+
+export const FATAL_LOCATION_DESTRUCTION: readonly CombatLocation[] = [
+  'head',
+  'center_torso',
+];
+
+export const STANDARD_STRUCTURE_TABLE: Readonly<
+  Record<
+    number,
+    {
+      head: number;
+      centerTorso: number;
+      sideTorso: number;
+      arm: number;
+      leg: number;
+    }
+  >
+> = {
+  20: { head: 3, centerTorso: 6, sideTorso: 5, arm: 3, leg: 4 },
+  25: { head: 3, centerTorso: 8, sideTorso: 6, arm: 4, leg: 6 },
+  30: { head: 3, centerTorso: 10, sideTorso: 7, arm: 5, leg: 7 },
+  35: { head: 3, centerTorso: 11, sideTorso: 8, arm: 6, leg: 8 },
+  40: { head: 3, centerTorso: 12, sideTorso: 10, arm: 6, leg: 10 },
+  45: { head: 3, centerTorso: 14, sideTorso: 11, arm: 7, leg: 11 },
+  50: { head: 3, centerTorso: 16, sideTorso: 12, arm: 8, leg: 12 },
+  55: { head: 3, centerTorso: 18, sideTorso: 13, arm: 9, leg: 13 },
+  60: { head: 3, centerTorso: 20, sideTorso: 14, arm: 10, leg: 14 },
+  65: { head: 3, centerTorso: 21, sideTorso: 15, arm: 10, leg: 15 },
+  70: { head: 3, centerTorso: 22, sideTorso: 15, arm: 11, leg: 15 },
+  75: { head: 3, centerTorso: 23, sideTorso: 16, arm: 12, leg: 16 },
+  80: { head: 3, centerTorso: 25, sideTorso: 17, arm: 13, leg: 17 },
+  85: { head: 3, centerTorso: 27, sideTorso: 18, arm: 14, leg: 18 },
+  90: { head: 3, centerTorso: 29, sideTorso: 19, arm: 15, leg: 19 },
+  95: { head: 3, centerTorso: 30, sideTorso: 20, arm: 16, leg: 20 },
+  100: { head: 3, centerTorso: 31, sideTorso: 21, arm: 17, leg: 21 },
+};
